Cardano.YTxP

Description

This module provides the blueprint for the Yielding Transaction Pattern Library (ytxp-lib).

Documentation

ytxpBlueprint :: Config -> SdkParameters -> ContractBlueprint

Generates the blueprint for the Yielding Transaction Pattern Library.

Since: 0.1.0