Cardano.YTxP.Control.Parameters

Description

Types

data YieldListScripts

Scripts that govern which transaction families can be "yielded to"

Since: 0.1.0

data YieldingScripts

Scripts that yield to transaction families described by the datums guarded by the YieldListScripts.

Since: 0.1.0

data ControlParameters

Contains the compiled scripts along with the parameters they were compiled against. This is useful for _library consumers_ and should contain all of the information needed to work with the library.

Since: 0.1.0

Parameter Generation (script compilation)

mkControlParameters :: ControlParametersInitial -> Either Text ControlParameters

Compile all scripts, threading through the appropriate parameters and script hashes