Cardano.YTxP.Control.ParametersInitial

Description

Documentation

data ControlParametersInitial

Parameters available to the YieldListValidator and YieldListMP during compilation (therefore not containing any script hashes).

This is a GADT because the nonces must be serializable (and thus haskell types) as well as able to be applied to plutarch scripts (and thus PTypes).

To load the scriptToWrap arguments, see unsafeTermFromScript from Cardano.YTxP.Control.Utils

Since: 0.1.0

Constructors

ControlParametersInitial

Fields

maxYieldListSize :: !Natural

If the yield list exceeds this size, blow up during STT minting @since 0.1.0

stakingValidatorsNonceList :: [Natural]

A list of nonces for the yielding staking validators. One staking validator is compiled for each nonce. @since 0.1.0

mintingPoliciesNonceList :: [Natural]

A list of nonces for the yielding minting policies. One minting policy is compiled for each nonce. @since 0.1.0

scriptToWrapYieldListMP :: forall (s :: S). Term s (PData :--> (PScriptContext :--> POpaque))

The V2 script that the yield list MP will wrap. This might be an admin signature script, multisig script, etc. @since 0.1.0

scriptToWrapYieldListValidator :: forall (s :: S). Term s (PData :--> (PData :--> (PScriptContext :--> POpaque)))

The V2 script that the yield list validator will wrap.

compilationConfig :: Config

Plutarch compilation config