Cardano.YTxP.Control.Stubs

Description

have some scripts fully implemented yet.

Documentation

alwaysSucceedsValidator :: forall (s :: S). Term s (PData :--> (PData :--> (PScriptContext :--> POpaque)))

A validator that will never fail

alwaysSucceedsTwoArgumentScript :: forall (s :: S). Term s (PData :--> (PScriptContext :--> POpaque))

A two argument script (minting policy or staking validator) that will never fail

noncedValidatorWrapper :: forall (nonceType :: S -> Type) (s :: S). Term s nonceType -> Term s (PData :--> (PData :--> (PScriptContext :--> POpaque))) -> Term s (PData :--> (PData :--> (PScriptContext :--> POpaque)))

Compile a nonce into the script, changing the script hash. This can be useful if you need many stubbed validators with different addresses.

noncedTwoArgumentScriptWrapper :: forall (nonceType :: S -> Type) (s :: S). Term s nonceType -> Term s (PData :--> (PScriptContext :--> POpaque)) -> Term s (PData :--> (PScriptContext :--> POpaque))

Compiles a nonce into the script, changing the script hashes. This is useful if you need many stubbed validators with different addresses