Cardano.YTxP.Control.Vendored

Description

Vendored utilities from open source libraries. See the appropriate License for details on usage.

Documentation

newtype DerivePConstantViaEnum (h :: Type) (p :: S -> Type)

Wrapper for deriving PConstantDecl using an Integer representation via Enum.

Vendored from LPE TODO: Licensing info

Since: 1.1.0

Constructors

DerivePConstantEnum h

newtype EnumIsData (a :: Type)

Wrapper for deriving ToData, FromData using an Integer representation via Enum.

Vendored from LPE TODO: Licensing info

Since: 1.1.0

Constructors

EnumIsData a

newtype ProductIsData (a :: Type)

Wrapper for deriving ToData, FromData using the List constructor of Data to represent a Product type.

It is recommended to use PlutusTypeDataList when deriving PlutusType as it provides some basic safety by ensuring Plutarch types have an Inner type of PDataRecord.

Uses gProductToBuiltinData, gproductFromBuiltinData.

Example

import qualified Generics.SOP as SOP

data Foo =
  Foo Integer [Integer]
  deriving stock (Generic)
  deriving anyclass (SOP.Generic)
  deriving (FromData, ToData) via (ProductIsData Foo)
  deriving (PConstantDecl) via (DerivePConstantViaDataList Foo PFoo)

instance PUnsafeLiftDecl PFoo where type PLifted PFoo = Foo

newtype PFoo s
    = PFoo
      ( Term s
          ( PDataRecord
              '[ "abc" ':= PInteger
               , "def" ':= PBuiltinList (PAsData PInteger)
               ]
          )
      )
  deriving stock (Generic)
  deriving anyclass (SOP.Generic)
  deriving anyclass (PlutusType, PIsData)

instance DerivePlutusType PFoo where
   type DPTStrat _ = PlutusTypeDataList

Since: 3.8.0

Constructors

ProductIsData a

unProductIsData :: forall (a :: Type). ProductIsData a -> a

Vendored from LPE @since 3.8.0

newtype DerivePConstantViaDataList (h :: Type) (p :: S -> Type)

Variant of PConstantViaData using the List repr from ProductIsData

Constructors

DerivePConstantViaDataList h

data PlutusTypeEnumData

PlutusTypeEnumData

Vendored from LPE TODO: Licensing info

data PlutusTypeDataList

Vendored from LPE @since 3.5.0

applyScript :: Script -> Script -> Script

Apply a function to an argument on the compiled Script level.

Vendored from LPE TODO: Licensing info

Since: 3.8.0

psymbolValueOf :: forall (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S). Term s (PCurrencySymbol :--> (PValue keys amounts :--> PInteger))

Get the sum of all values belonging to a particular CurrencySymbol. Modified version of LPE function with same purpose.