Cardano.YTxP.Control.YieldList

Description

Hashes

data YieldedToHash

A single hash that a yielding script can yield to A yielded to script can be a validator, minting policy or a stake validator

Constructors

YieldedToValidator CustomScriptHash

YieldedToMP CustomScriptHash

YieldedToSV CustomScriptHash

data CustomScriptHash

We use this CustomScriptHash instead of ScriptHash in order to ensure that the hash is of length 28.

tryMkCustomScriptHash :: BuiltinByteString -> CustomScriptHash

Note(Nigel): This will likely not compile under `plutus-tx` due to the use of error from the Haskell Prelude. We use error from Prelude here as using traceError doesn't give back the error message. See the following issue for more details: https://github.com/IntersectMBO/plutus/issues/3003

data PYieldedToHash (s :: S)

Constructors

PYieldedToValidator (Term s (PDataRecord '["scriptHash" ':= PScriptHash]))

PYieldedToMP (Term s (PDataRecord '["scriptHash" ':= PScriptHash]))

PYieldedToSV (Term s (PDataRecord '["scriptHash" ':= PScriptHash]))

Redeemers

data YieldListMPWrapperRedeemer

Redeemer for mkYieldListMPWrapper.

data PYieldListMPWrapperRedeemer (s :: S)

Plutarch-level version of YieldListMPWrapperRedeemer.

Constructors

PMint

PBurn

Datums

newtype YieldListDatum

The YieldListDatum holds a collection of hashes that YieldingScripts can yield to. The length of the datum is checked upon creation in mkYieldListSTMPWrapper to ensure that the length of the list does not exceed the max list length passed as a parameter to that script.

Constructors

YieldListDatum [YieldedToHash]

newtype PYieldListDatum (s :: S)

Constructors

PYieldListDatum (Term s (PDataRecord '["yieldedToScripts" ':= PBuiltinList (PAsData PYieldedToHash)]))

Functions

getYieldedToHashByIndex :: Term s (PYieldListDatum :--> (PInteger :--> PYieldedToHash))