Cardano.YTxP.Control.Yielding

Documentation

data YieldListIndex

Represents an index into a YieldList

data YieldListRefInputIndex

Represents an index into reference inputs of a transaction. The UTxO at this index must contain a YieldListSTT; otherwise, we blow up

data YieldingRedeemer

The redeemer passed to the yielding minting policy, validator, and staking validators

Constructors

YieldingRedeemer YieldListIndex YieldListScriptToYieldIndex YieldListRefInputIndex

getYieldedToHash :: YieldListSTCS -> Term s (PBuiltinList PTxInInfo :--> (PYieldingRedeemer :--> PYieldedToHash))

Given a list of reference inputs and a Yielding Redeemer, dig out the YieldList by: - Indexing the reference inputs according to the redeemer - Checking the fetched reference input for the correct YieldListSTCS - Decoding its datum (unsafely; the presence of the YieldListSTCS ensure it is authentic and well-formed) - Looking in the datum at the index in the redeemer and returning the YieldedToHash