Cardano.YTxP.Control.Yielding.StakingValidator

Documentation

compileYieldingSV :: Config -> YieldListSTCS -> Natural -> Either Text (HexStringScript "YieldingSV")

Compile a yielding staking validator that has been nonced. The nonce is required because each staking validator can only be delegated to a single pool; the inclusion of the nonce will change the script hash.