Utils

Documentation

pmintFieldHasTokenOfCurrencySymbolTokenNameAndAmount :: forall (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S). Term s (PValue keys amounts :--> (PCurrencySymbol :--> (PTokenName :--> (PInteger :--> PBool))))

Check that there is only token of given PCurrencySymbol and PTokenName with given amount contained in the given PValue.

Note: We use this function when checking the mint field of TxInfo, as when checking mint we need to check that the length is equal to two to account for the additional 'zero Ada' entry automatically included in the mint field: (PCurrencySymbol 0x,PMap [(PTokenName 0x,0)])])

phasOneScriptInputAtValidatorWithExactlyOneToken :: Term s (PBuiltinList PTxInInfo) -> Term s (PCurrencySymbol :--> (PTxOutRef :--> PBool))

phasOnlyOneValidScriptOutputWithToken :: Natural -> Term s (PBuiltinList PTxOut) -> Term s (PCurrencySymbol :--> PBool)

Check that: - there is exactly one script output with exactly one token with the given PCurrencySymbol and PTokenName - there is no other script output with one or more of these tokens (We do this in order to use this helper to ensure that there are no other script outputs containing one ore more of this token) - there is no wallet output that contains one of these tokens - if there is exactly one script output with exactly one of these tokens, we check that it contains no other tokens aside from Ada - finally check that the output also contains a valid (according to the spec) POutputDatum, ensuring the length of the yield list does not exceed the provided maxYieldListSize parameter.

phasOnlyOneInputWithExactlyOneTokenWithSymbol :: Term s (PBuiltinList PTxInInfo) -> Term s (PCurrencySymbol :--> PBool)

Check that there is exactly one input, and that one input carries exactly one yield list token.

phasNoScriptInputWithToken :: Term s (PBuiltinList PTxInInfo) -> Term s (PCurrencySymbol :--> PBool)

poutputsDoNotContainToken :: Term s (PBuiltinList PTxOut) -> Term s (PCurrencySymbol :--> PBool)

Check that none of the given outputs contain a token with the given PCurrencySymbol

pemptyTokenName :: Term s PTokenName

Empty token name

pands :: [Term s PBool] -> Term s PBool

Like Haskell's and but for Plutarch terms Bool has the same function but does not export it.

pscriptHashToCurrencySymbol :: Term s PScriptHash -> Term s PCurrencySymbol

Convert a ScriptHash to a CurrencySymbol, which has the same representation

punsafeFromInlineDatum :: forall (s :: S) (a :: S -> Type). Term s (POutputDatum :--> a)

Extract the datum from a POutputDatum, expecting it to be an inline datum.

pmember :: PIsData k => Term s (k :--> (PMap any k v :--> PBool))