Cardano.YTxP.Control.Yielding

Description

Add note on orphan instances. We want to have the haskell types come from the shared SDK so that other onchain implementations can reuse those same types. This however, forces us to declare orphan instances for haskell -> plutarch conversion

Documentation

getAuthorisedScriptHash :: forall (s :: S). Term s (PCurrencySymbol :--> (PBuiltinList PTxInInfo :--> (PYieldingRedeemer :--> PScriptHash)))

Given a list of reference inputs and a Yielding Redeemer, dig out the authorised script hash by:

Indexing the reference inputs according to the redeemer

Checking the fetched reference input for the correct AuthorisedScriptsSTCS

Returning the AuthorisedScriptHash

data PAuthorisedScriptPurpose (s :: S)

Constructors

PMinting

PSpending

PRewarding

data PYieldingRedeemer (s :: S)

Orphan instances

PConstantDecl AuthorisedScriptIndex

Instance details

Associated Types

type PConstantRepr AuthorisedScriptIndex

type PConstanted AuthorisedScriptIndex :: PType

Methods

pconstantToRepr :: AuthorisedScriptIndex -> PConstantRepr AuthorisedScriptIndex

pconstantFromRepr :: PConstantRepr AuthorisedScriptIndex -> Maybe AuthorisedScriptIndex

PConstantDecl AuthorisedScriptProofIndex

Instance details

Associated Types

type PConstantRepr AuthorisedScriptProofIndex

type PConstanted AuthorisedScriptProofIndex :: PType

Methods

pconstantToRepr :: AuthorisedScriptProofIndex -> PConstantRepr AuthorisedScriptProofIndex

pconstantFromRepr :: PConstantRepr AuthorisedScriptProofIndex -> Maybe AuthorisedScriptProofIndex

PConstantDecl AuthorisedScriptPurpose

Instance details

Associated Types

type PConstantRepr AuthorisedScriptPurpose

type PConstanted AuthorisedScriptPurpose :: PType

Methods

pconstantToRepr :: AuthorisedScriptPurpose -> PConstantRepr AuthorisedScriptPurpose

pconstantFromRepr :: PConstantRepr AuthorisedScriptPurpose -> Maybe AuthorisedScriptPurpose

PConstantDecl YieldingRedeemer

Instance details

Associated Types

type PConstantRepr YieldingRedeemer

type PConstanted YieldingRedeemer :: PType

Methods

pconstantToRepr :: YieldingRedeemer -> PConstantRepr YieldingRedeemer

pconstantFromRepr :: PConstantRepr YieldingRedeemer -> Maybe YieldingRedeemer