Cardano.YTxP.Control.Yielding.StakingValidator

Staking Validator

data YieldingStakingValidatorScript

A yielding staking validator together with its nonce.

Since: 0.1.0

compileYieldingStakingValidator :: Config -> YieldListSTCS -> Natural -> Either Text YieldingStakingValidatorScript

Compile a yielding staking validator that has been nonced. The nonce is required because each staking validator can only be delegated to a single pool; the inclusion of the nonce will change the script hash.

Credential and Nonce